+++ mitra_cli/src/cli.rs
@@ -729,8 +729,7 @@ impl ListUnreachableActors {
         for profile in profiles {
             println!(
                 "{0: <60} | {1: <35} | {2: <35}",
-                profile.actor_id
-                    .expect("actor ID should be present"),
+                profile.expect_remote_actor_id(),
                 profile.unreachable_since
                     .expect("unreachable flag should be present")
                     .to_string(),