+++ src/job_queue/periodic_tasks.rs
@@ -148,10 +148,7 @@ pub async fn delete_empty_profiles(
         let profile = get_profile_by_id(db_client, &profile_id).await?;
         let deletion_queue = delete_profile(db_client, &profile.id).await?;
         delete_media(config, deletion_queue).await;
-        log::info!(
-            "deleted empty profile {}",
-            profile.expect_remote_actor_id(),
-        );
+        log::info!("deleted empty profile {}", profile);
     };
     Ok(())
 }