+++ src/mastodon_api/oauth/views.rs
@@ -241,7 +241,7 @@ async fn token_view(
         created_at,
         expires_at,
     ).await?;
-    log::warn!("created auth token for user {}", user.id);
+    log::warn!("created auth token for user {}", user);
     let token_response = TokenResponse::new(
         access_token,
         created_at.timestamp(),