+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Increased Monero Wallet RPC timeout to 15 seconds.
+- Use actor ID or username in log messages instead of profile UUID.
 
 ## [2.14.0] - 2024-03-30